refactor: move assessment and submission info to separate handlers
